+++ External/FEXCore/Source/CMakeLists.txt
@@ -1,7 +1,6 @@
 set (MAN_DIR ${CMAKE_INSTALL_PREFIX}/share/man CACHE PATH "MAN_DIR")
 
 set (FEXCORE_BASE_SRCS
-  Common/Paths.cpp
   Interface/Config/Config.cpp
   Utils/Allocator.cpp
   Utils/Allocator/64BitAllocator.cpp
