--- External/FEXCore/Source/Common/Paths.h
@@ -1,12 +1,0 @@
-#pragma once
-#include <FEXCore/fextl/string.h>
-
-namespace FEXCore::Paths {
-  void InitializePaths();
-  void ShutdownPaths();
-
-  const char *GetHomeDirectory();
-
-  fextl::string GetCachePath();
-  fextl::string GetEntryCachePath();
-}
